Make Quota and Redundancy uint64

+++ providers/clusterconf/bundle.go
@@ -33,7 +33,7 @@ type Bundle struct {
 	ID         uint64                   `json:"id"`
 	Datasets   map[string]BundleDataset `json:"datasets"`
 	Services   map[string]BundleService `json:"services"`
-	Redundancy int                      `json:"redundancy"`
+	Redundancy uint64                   `json:"redundancy"`
 	Ports      BundlePorts              `json:"ports"`
 	// ModIndex should be treated as opaque, but passed back on updates.
 	ModIndex uint64 `json:"modIndex"`
@@ -76,7 +76,7 @@ type BundleDataset struct {
 	Name  string            `json:"name"`
 	ID    string            `json:"id"`
 	Type  BundleDatasetType `json:"type"`
-	Quota int               `json:"type"`
+	Quota uint64            `json:"type"`
 }
 
 func (d BundleDataset) overlayOn(base *Dataset) (BundleDataset, error) {
